Chore (cleaning):
moving functions to amore logical file structure
